requirements.txt
certifi==2019.11.28
chardet==3.0.4
dictor==0.1.5
idna==2.9
PyYAML==5.3
requests==2.23.0
urllib3==1.25.8
